+++ node/cmd/guardiand/adminnodes.go
@@ -118,6 +118,9 @@ func runListNodes(cmd *cobra.Command, args []string) {
 		{"Mantle", vaa.ChainIDMantle},
 		{"Wormchain", vaa.ChainIDWormchain},
 		{"Sepolia", vaa.ChainIDSepolia},
+		{"ArbitrumSepolia", vaa.ChainIDArbitrumSepolia},
+		{"BaseSepolia", vaa.ChainIDBaseSepolia},
+		{"OptimismSepolia", vaa.ChainIDOptimismSepolia},
 	}
 
 	if len(only) > 0 {