+++ node/cmd/guardiand/node.go
@@ -181,6 +181,15 @@ var (
 	sepoliaRPC      *string
 	sepoliaContract *string
 
+	arbitrumSepoliaRPC      *string
+	arbitrumSepoliaContract *string
+
+	baseSepoliaRPC      *string
+	baseSepoliaContract *string
+
+	optimismSepoliaRPC      *string
+	optimismSepoliaContract *string
+
 	logLevel                *string
 	publicRpcLogDetailStr   *string
 	publicRpcLogToTelemetry *bool
@@ -352,6 +361,15 @@ func init() {
 	baseRPC = NodeCmd.Flags().String("baseRPC", "", "Base RPC URL")
 	baseContract = NodeCmd.Flags().String("baseContract", "", "Base contract address")
 
+	arbitrumSepoliaRPC = NodeCmd.Flags().String("arbitrumSepoliaRPC", "", "Arbitrum on Sepolia RPC URL")
+	arbitrumSepoliaContract = NodeCmd.Flags().String("arbitrumSepoliaContract", "", "Arbitrum on Sepolia contract address")
+
+	baseSepoliaRPC = NodeCmd.Flags().String("baseSepoliaRPC", "", "Base on Sepolia RPC URL")
+	baseSepoliaContract = NodeCmd.Flags().String("baseSepoliaContract", "", "Base on Sepolia contract address")
+
+	optimismSepoliaRPC = NodeCmd.Flags().String("optimismSepoliaRPC", "", "Optimism on Sepolia RPC URL")
+	optimismSepoliaContract = NodeCmd.Flags().String("optimismSepoliaContract", "", "Optimism on Sepolia contract address")
+
 	logLevel = NodeCmd.Flags().String("logLevel", "info", "Logging level (debug, info, warn, error, dpanic, panic, fatal)")
 	publicRpcLogDetailStr = NodeCmd.Flags().String("publicRpcLogDetail", "full", "The detail with which public RPC requests shall be logged (none=no logging, minimal=only log gRPC methods, full=log gRPC method, payload (up to 200 bytes) and user agent (up to 200 bytes))")
 	publicRpcLogToTelemetry = NodeCmd.Flags().Bool("logPublicRpcToTelemetry", true, "whether or not to include publicRpc request logs in telemetry")
@@ -505,6 +523,9 @@ func runNode(cmd *cobra.Command, args []string) {
 		*sepoliaContract = unsafeDevModeEvmContractAddress(*sepoliaContract)
 		*scrollContract = unsafeDevModeEvmContractAddress(*scrollContract)
 		*mantleContract = unsafeDevModeEvmContractAddress(*mantleContract)
+		*arbitrumSepoliaContract = unsafeDevModeEvmContractAddress(*arbitrumSepoliaContract)
+		*baseSepoliaContract = unsafeDevModeEvmContractAddress(*baseSepoliaContract)
+		*optimismSepoliaContract = unsafeDevModeEvmContractAddress(*optimismSepoliaContract)
 	}
 
 	// Verify flags
@@ -665,31 +686,38 @@ func runNode(cmd *cobra.Command, args []string) {
 		logger.Fatal("If --gatewayWS is not specified, then --gatewayLCD and --gatewayContract must not be specified")
 	}
 
-	if *testnetMode {
-		if *neonRPC == "" {
-			logger.Fatal("Please specify --neonRPC")
+	// These chains are only allowed in devnet and testnet.
+	if *testnetMode || *unsafeDevMode {
+		if (*neonRPC == "") != (*neonContract == "") {
+			logger.Fatal("Both --neonRPC and --neonContract must be set together or both unset")
 		}
-		if *neonContract == "" {
-			logger.Fatal("Please specify --neonContract")
+		if (*sepoliaRPC == "") != (*sepoliaContract == "") {
+			logger.Fatal("Both --sepoliaRPC and --sepoliaContract must be set together or both unset")
 		}
-		if *sepoliaRPC == "" {
-			logger.Fatal("Please specify --sepoliaRPC")
+		if (*arbitrumSepoliaRPC == "") != (*arbitrumSepoliaContract == "") {
+			logger.Fatal("Both --arbitrumSepoliaRPC and --arbitrumSepoliaContract must be set together or both unset")
 		}
-		if *sepoliaContract == "" {
-			logger.Fatal("Please specify --sepoliaContract")
+		if (*baseSepoliaRPC == "") != (*baseSepoliaContract == "") {
+			logger.Fatal("Both --baseSepoliaRPC and --baseSepoliaContract must be set together or both unset")
+		}
+		if (*optimismSepoliaRPC == "") != (*optimismSepoliaContract == "") {
+			logger.Fatal("Both --optimismSepoliaRPC and --optimismSepoliaContract must be set together or both unset")
 		}
 	} else {
-		if *neonRPC != "" && !*unsafeDevMode {
-			logger.Fatal("Please do not specify --neonRPC")
+		if *neonRPC != "" || *neonContract != "" {
+			logger.Fatal("Please do not specify --neonRPC or --neonContract")
 		}
-		if *neonContract != "" && !*unsafeDevMode {
-			logger.Fatal("Please do not specify --neonContract")
+		if *sepoliaRPC != "" || *sepoliaContract != "" {
+			logger.Fatal("Please do not specify --sepoliaRPC or --sepoliaContract")
 		}
-		if *sepoliaRPC != "" && !*unsafeDevMode {
-			logger.Fatal("Please do not specify --sepoliaRPC")
+		if *arbitrumSepoliaRPC != "" || *arbitrumSepoliaContract != "" {
+			logger.Fatal("Please do not specify --arbitrumSepoliaRPC or --arbitrumSepoliaContract")
 		}
-		if *sepoliaContract != "" && !*unsafeDevMode {
-			logger.Fatal("Please do not specify --sepoliaContract")
+		if *baseSepoliaRPC != "" || *baseSepoliaContract != "" {
+			logger.Fatal("Please do not specify --baseSepoliaRPC or --baseSepoliaContract")
+		}
+		if *optimismSepoliaRPC != "" || *optimismSepoliaContract != "" {
+			logger.Fatal("Please do not specify --optimismSepoliaRPC or --optimismSepoliaContract")
 		}
 	}
 
@@ -893,6 +921,9 @@ func runNode(cmd *cobra.Command, args []string) {
 	rpcMap["pythnetWS"] = *pythnetWS
 	if env == common.TestNet {
 		rpcMap["sepoliaRPC"] = *sepoliaRPC
+		rpcMap["arbitrumSepoliaRPC"] = *arbitrumSepoliaRPC
+		rpcMap["baseSepoliaRPC"] = *baseSepoliaRPC
+		rpcMap["optimismSepoliaRPC"] = *optimismSepoliaRPC
 	}
 	rpcMap["scrollRPC"] = *scrollRPC
 	rpcMap["solanaRPC"] = *solanaRPC
@@ -1428,28 +1459,61 @@ func runNode(cmd *cobra.Command, args []string) {
 		watcherConfigs = append(watcherConfigs, wc)
 	}
 
-	if *testnetMode {
-		if shouldStart(neonRPC) {
-			if !shouldStart(solanaRPC) {
-				log.Fatalf("If neon is enabled then solana must also be enabled.")
-			}
-			wc := &evm.WatcherConfig{
-				NetworkID:           "neon",
-				ChainID:             vaa.ChainIDNeon,
-				Rpc:                 *neonRPC,
-				Contract:            *neonContract,
-				L1FinalizerRequired: "solana-finalized",
-			}
-
-			watcherConfigs = append(watcherConfigs, wc)
+	if *testnetMode && shouldStart(neonRPC) {
+		if !shouldStart(solanaRPC) {
+			log.Fatalf("If neon is enabled then solana must also be enabled.")
+		}
+		wc := &evm.WatcherConfig{
+			NetworkID:           "neon",
+			ChainID:             vaa.ChainIDNeon,
+			Rpc:                 *neonRPC,
+			Contract:            *neonContract,
+			L1FinalizerRequired: "solana-finalized",
 		}
 
+		watcherConfigs = append(watcherConfigs, wc)
+	}
+
+	if *testnetMode || *unsafeDevMode {
 		if shouldStart(sepoliaRPC) {
 			wc := &evm.WatcherConfig{
 				NetworkID: "sepolia",
 				ChainID:   vaa.ChainIDSepolia,
 				Rpc:       *sepoliaRPC,
 				Contract:  *sepoliaContract,
+			}
+
+			watcherConfigs = append(watcherConfigs, wc)
+		}
+
+		if shouldStart(arbitrumSepoliaRPC) {
+			wc := &evm.WatcherConfig{
+				NetworkID: "arbitrum_sepolia",
+				ChainID:   vaa.ChainIDArbitrumSepolia,
+				Rpc:       *arbitrumSepoliaRPC,
+				Contract:  *arbitrumSepoliaContract,
+			}
+
+			watcherConfigs = append(watcherConfigs, wc)
+		}
+
+		if shouldStart(baseSepoliaRPC) {
+			wc := &evm.WatcherConfig{
+				NetworkID: "base_sepolia",
+				ChainID:   vaa.ChainIDBaseSepolia,
+				Rpc:       *baseSepoliaRPC,
+				Contract:  *baseSepoliaContract,
+			}
+
+			watcherConfigs = append(watcherConfigs, wc)
+		}
+
+		if shouldStart(optimismSepoliaRPC) {
+			wc := &evm.WatcherConfig{
+				NetworkID: "optimism_sepolia",
+				ChainID:   vaa.ChainIDOptimismSepolia,
+				Rpc:       *optimismSepoliaRPC,
+				Contract:  *optimismSepoliaContract,
 			}
 
 			watcherConfigs = append(watcherConfigs, wc)