+++ sdk/js/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.10.7
+
+### Added
+
+Arbitrum on Sepolia support
+
+Base on Sepolia support
+
+Optimism on Sepolia support
+
 ## 0.10.6
 
 ### Added